refactor: split flow_service.py into modular mixins
Split flow_service.py (566 lines) into 3 specialized mixins:
- flow_read_mixin.py (135 lines): read operations (get_flow_state, get_flow_status, list_flows, get_flow_timeline)
- flow_write_mixin.py (169 lines): write operations (create_flow, update_flow_metadata, delete_flow, bind_spec)
- flow_transition.py (260 lines): transition operations (ensure_flow_for_branch, reactivate_flow)

Architecture:
- FlowService inherits FlowLifecycleMixin + FlowTransitionMixin
- FlowTransitionMixin inherits FlowWriteMixin + FlowReadMixin
- FlowWriteMixin inherits FlowReadMixin
- Inheritance chain ensures all methods available to FlowService

Type safety:
- Use Self type annotation for mixin methods
- Add cast() for git_client return values (Any → str)
- MyPy strict mode passes

Tests:
- Update mock paths: GitHubClient → flow_read_mixin, SignatureService → flow_write_mixin
- All 24 flow tests pass

Refs: #3056, #3057
