+++ tests/vibe3/services/test_flow_events.py
@@ -13,11 +13,11 @@ from vibe3.services.flow_service import FlowService
 def stable_worktree_actor(monkeypatch):
     """Avoid real git identity lookups during flow creation tests."""
     monkeypatch.setattr(
-        "vibe3.services.flow_service.SignatureService.get_worktree_actor",
+        "vibe3.services.flow_write_mixin.SignatureService.get_worktree_actor",
         lambda: "test-actor",
     )
     monkeypatch.setattr(
-        "vibe3.services.flow_service.GitHubClient.get_pr",
+        "vibe3.services.flow_read_mixin.GitHubClient.get_pr",
         lambda self, pr_number=None, branch=None: None,
     )
