+++ tests/vibe3/services/test_task_service_fresh_db.py
@@ -14,7 +14,7 @@ def stable_flow_actor(monkeypatch):
         lambda store, branch, explicit_actor=None: explicit_actor or "test-actor",
     )
     monkeypatch.setattr(
-        "vibe3.services.flow_service.GitHubClient.get_pr",
+        "vibe3.services.flow_read_mixin.GitHubClient.get_pr",
         lambda self, pr_number=None, branch=None: None,
     )
